refactor: split app.py into secrets, upstream, and streaming modules (plan 002)
